refactor: 지원서 페이지 Server/Client Component 분리 (#95)
